score_service/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
fastapi==0.115.6
grpcio==1.68.1
h11==0.14.0
idna==3.10
protobuf==5.29.1
pydantic==2.10.3
pydantic_core==2.27.1
python-multipart==0.0.19
requests==2.32.3
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
